feat: Centralize database connection management and config loading
